Finished tutorial level and moved movement files -movement files don't work -added door and door scripts -keys can now be picked up
