+++ Assets/Scripts/CharacterCollisions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87544a40d93371f439830ff547b70e1d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
