+++ Assets/Scripts/Collectible.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5ae1f16750132448a645ee19389128d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
