+++ Assets/Scripts/Collectible/DoorScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2eb2e3fcd1f7df54cb1b50ecf1cc0585
+guid: d67b09636708c2049b7f442f877c3d8c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
