+++ Assets/Scripts/Collectible/LevelTransition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7fb73a71452ba4e46a6bc0bc773392d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
