+++ Assets/Scripts/Collectible/PickupFloatAnimation.cs
