+++ Assets/Scripts/PlatformSwitch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 175fdfa9855b0d248a500e9c02800257
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
