+++ Assets/Shaders/Subgraphs.meta~merged.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 86bfbb213e0f07449ace146fd1aa58ca
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
