--- PlayerController.cs
@@ -1,96 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PlayerController : MonoBehaviour
-{
-
-    Rigidbody rb;
-    public bool isGrounded;
-
-    public float velocity = 5;
-    public float turnspeed = 10;
-    public float jumpSpeed = 5f;
-    public float fallMultiplier = 2.5f;
-    public float lowJumpMultiplier = 2f;
-
-    Vector2 input;
-    float angle;
-
-    Quaternion targetRotation;
-    Transform cam;
-
-    void AWake()
-    {
-        rb = GetComponent<Rigidbody>();
-    }
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        cam = Camera.main.transform;
-        rb = GetComponent<Rigidbody>();
-    }
-
-    void OnCollisionStay()
-    {
-        isGrounded = true;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        GetInput();
-
-        if (Mathf.Abs(input.x) < 1 && Mathf.Abs(input.y) < 1)
-        {
-            return;
-        }
-        CalculateDirection();
-        Rotate();
-        Move();
-
-        if (Input.GetKeyDown(KeyCode.Space) && isGrounded)
-        {
-            rb.AddForce(new Vector3(0, 5, 0), ForceMode.Impulse);
-            isGrounded = false;
-        }
-
-        if(rb.velocity.y < 0)
-        {
-            rb.velocity += Vector3.up * Physics.gravity.y * (fallMultiplier - 1) * Time.deltaTime;
-        }
-
-        else if (rb.velocity.y > 0 && Input.GetKeyDown(KeyCode.Space))
-        {
-            rb.velocity += Vector3.up * Physics.gravity.y * (lowJumpMultiplier - 1) * Time.deltaTime;
-        }
-
-    }
-
-    void GetInput()
-    {
-        input.x = Input.GetAxisRaw("Horizontal");
-        input.y = Input.GetAxisRaw("Vertical");
-    }
-
-    void CalculateDirection()
-    {
-        angle = Mathf.Atan2(input.x, input.y);
-        angle = Mathf.Rad2Deg * angle;
-        angle += cam.eulerAngles.y;
-
-    }
-
-    void Rotate()
-    {
-        targetRotation = Quaternion.Euler(0, angle, 0);
-        transform.rotation = Quaternion.Slerp(transform.rotation, targetRotation, turnspeed * Time.deltaTime);
-    }
-
-    void Move()
-    {
-        transform.position += transform.forward * velocity * Time.deltaTime;
-    }
-
-}
