+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,9 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/SampleScene.unity
-    guid: d1c3109bdb54ad54c8a2b2838528e640
+    path: Assets/Scenes/Prototypes/TutorialPrototype.unity
+    guid: 34a21ccfb3f05854ba4a86fe32a90cf9
+  - enabled: 1
+    path: Assets/Scenes/Levels/FirstLevel.unity
+    guid: 826d3e768c26b4a4a85825eb7f4d7cdd
   m_configObjects: {}
